Some type fixing
