+++ composer.json
@@ -23,6 +23,7 @@
         "symfony/messenger": "^4.2 || ^5.0",
         "symfony/security-guard": "^4.0 || ^5.0",
         "symfony/validator": "*",
+        "twig/twig": "^2.10 || ^3.0",
         "weirdan/codeception-psalm-module": "~0.8"
     },
     "autoload": {
