+++ src/Twig/TemplateFileAnalyzer.php
@@ -5,25 +5,12 @@ declare(strict_types=1);
 namespace Psalm\SymfonyPsalmPlugin\Twig;
 
 
-use Psalm\CodeLocation;
 use Psalm\Context;
 use Psalm\Internal\Analyzer\FileAnalyzer;
 use Psalm\Internal\Codebase\Taint;
-use Psalm\Internal\Taint\Sink;
-use Psalm\Internal\Taint\Taintable;
 use Psalm\Internal\Taint\TaintNode;
-use Psalm\Type;
-use Psalm\Type\TaintKind;
 use Twig\Environment;
 use Twig\Loader\FilesystemLoader;
-use Twig\Node\BodyNode;
-use Twig\Node\Expression\AbstractExpression;
-use Twig\Node\Expression\FilterExpression;
-use Twig\Node\Expression\NameExpression;
-use Twig\Node\ModuleNode;
-use Twig\Node\Node;
-use Twig\Node\PrintNode;
-use Twig\Source;
 
 class TemplateFileAnalyzer extends FileAnalyzer
 {
@@ -31,11 +18,6 @@ class TemplateFileAnalyzer extends FileAnalyzer
      * @var Taint|null
      */
     private $taint;
-
-    /**
-     * @var Source
-     */
-    private $twig_source;
 
     public function analyze(
         Context $file_context = null,
@@ -60,14 +42,16 @@ class TemplateFileAnalyzer extends FileAnalyzer
         ]);
 
         $local_file_name = str_replace('templates/', '', $this->file_name);
-        $this->twig_source = $loader->getSourceContext($local_file_name);
-        $tree = $twig->parse($twig->tokenize($this->twig_source));
+        $twig_source = $loader->getSourceContext($local_file_name);
+        $tree = $twig->parse($twig->tokenize($twig_source));
 
-        $this->context = new Context();
+        $context = new Context();
 
-        $this->analyzeModule($tree);
-        foreach ($this->context->vars_in_scope as $var_name => $var_type) {
-            $taint_source = self::getTaintNodeForTwigNamedVariable($local_file_name, $var_name, null);
+        $moduleAnalyzer = new TwigModuleAnalyzer($context, $twig_source, $codebase->taint);
+        $moduleAnalyzer->analyzeModule($tree);
+
+        foreach ($context->vars_in_scope as $var_name => $var_type) {
+            $taint_source = self::getTaintNodeForTwigNamedVariable($local_file_name, $var_name);
             $this->taint->addTaintNode($taint_source);
 
             if ($var_type->parent_nodes === null) {
@@ -78,128 +62,6 @@ class TemplateFileAnalyzer extends FileAnalyzer
                 $this->taint->addPath($taint_source, $taint_sink, 'arg');
             }
         }
-    }
-
-    private function analyzeModule(ModuleNode $node): void
-    {
-        $this->analyzeBody($node->getNode('body'));
-    }
-
-    private function analyzeBody(BodyNode $node): void
-    {
-        $children = $node->getIterator()[0]->getIterator();
-        /** @var Node $sub_node */
-        while($sub_node = $children->current()) {
-            if ($sub_node instanceof PrintNode) {
-                $this->analyzePrintNode($sub_node);
-            }
-            $children->next();
-        }
-    }
-
-    private function analyzePrintNode(PrintNode $node): Taintable
-    {
-        $codeLocation = self::getLocation($node->getSourceContext() ?? $this->twig_source, $node->getTemplateLine());
-
-        $sink = Sink::getForMethodArgument(
-            'twig_print', 'twig_print', 0, null, $codeLocation
-        );
-
-        $sink->taints = [
-            TaintKind::INPUT_HTML,
-            TaintKind::USER_SECRET,
-            TaintKind::SYSTEM_SECRET
-        ];
-
-        $this->taint->addSink($sink);
-
-        $expression = $node->getNode('expr');
-        if($expression_taint = $this->analyzeExpression($expression)) {
-            $this->taint->addPath($expression_taint, $sink, 'arg');
-        }
-
-        return $sink;
-    }
-
-    private function analyzeExpression(AbstractExpression $expression): ?Taintable
-    {
-        if ($expression instanceof FilterExpression) {
-            return $this->analyzeFilter($expression);
-        } else {
-            return null;
-        }
-    }
-
-    private function analyzeFilter(FilterExpression $expression): ?Taintable
-    {
-        if ($expression->getNode('filter')->getAttribute('value') !== 'raw') {
-            return null;
-        }
-
-        $function_id = 'filter_raw';
-        $return_location = self::getLocation($expression->getSourceContext(), $expression->getTemplateLine());
-        $return_taint = TaintNode::getForMethodReturn($function_id, $function_id, $return_location, $return_location);
-
-        $this->taint->addTaintNode($return_taint);
-
-        $argument_taint = TaintNode::getForMethodArgument(
-            $function_id, $function_id, 0, $return_location, $return_location // should be $argument_location instead of $return_location
-        );
-
-        $this->taint->addTaintNode($argument_taint);
-        $this->taint->addPath($argument_taint, $return_taint, 'arg');
-
-        $sub_node = $expression->getNode('node');
-        if ($sub_node instanceof NameExpression) {
-            $sub_node_taint = $this->analyzeName($sub_node);
-        }
-
-        $this->taint->addPath($sub_node_taint, $argument_taint, 'arg');
-
-        return $return_taint;
-    }
-
-    private function analyzeName(NameExpression $expression): Taintable
-    {
-        $var_id = $expression->getAttribute('name');
-        $var_type = Type::getMixed();
-        $this->context->vars_in_scope[$var_id] = $var_type;
-
-        $var_location = self::getLocation($expression->getSourceContext(), $expression->getTemplateLine());
-        $variable_taint = TaintNode::getForAssignment($var_id, $var_location);
-        $this->taint->addTaintNode($variable_taint);
-
-        $var_type->parent_nodes = [$variable_taint];
-
-        return $variable_taint;
-    }
-
-    private static function getLocation(Source $sourceContext, int $lineNumber) : CodeLocation
-    {
-        $fileName = $sourceContext->getName();
-        $filePath = $sourceContext->getPath();
-        $snippet = (string) $sourceContext->getCode(); // warning : the getCode method returns the whole template, not only the statement
-        $fileCode = file_get_contents($filePath);
-
-        $lines = explode("\n", $snippet);
-
-        $file_start = 0;
-
-        for ($i = 0; $i < $lineNumber - 1; $i++) {
-            $file_start += strlen($lines[$i]) + 1;
-        }
-
-        $lineNumber = $lineNumber ?: 1;
-        $file_start += strpos($lines[$lineNumber - 1], $snippet);
-        $file_end = $file_start + strlen($snippet);
-
-        return new CodeLocation\Raw(
-            $fileCode,
-            $filePath,
-            $fileName,
-            $file_start,
-            max($file_end, strlen($fileCode))
-        );
     }
 
     public static function getTaintNodeForTwigNamedVariable(
